backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
django-extensions==3.2.3
Pillow==10.0.1
channels==4.0.0
channels-redis==4.1.0
uvicorn==0.24.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
whitenoise==6.6.0
